Separate screens to own modules
