+++ App.tsx
@@ -1,11 +1,13 @@
-import React, { useState } from 'react';
-import { Text, View, StatusBar, Button } from 'react-native';
+import React from 'react';
 import { createAppContainer, createStackNavigator } from 'react-navigation';
-import styled from 'styled-components';
-
-import { connect, Provider } from 'react-redux'
+import { Provider } from 'react-redux'
 import { createStore } from 'redux'
+
 import countReducer from './reducers/count'
+
+import CountScreen from './screens/CountScreen'
+import HomeScreen from './screens/HomeScreen'
+import TimeScreen from './screens/TimeScreen'
 
 const store = createStore(countReducer)
 
@@ -22,61 +24,6 @@ class App extends React.Component<{}, AppState> {
       </Provider>
     )
   }
-}
-
-const Container = styled(View)`
-  align-items: center;
-  background-color: pink;
-  display: flex;
-  justify-content: center;
-  height: 100%;
-  width: 100%;
-`
-
-const CountScreen = () => {
-
-  const [count, setCount] = useState(0)
-
-  return(
-  <Container>
-    <StatusBar hidden={true}/>
-    <Text>You clicked {count} times. How fun.</Text>
-    <Button
-      title="Wheeee"
-      onPress={() => setCount(count + 1)}
-    />
-  </Container>
-  )
-}
-
-const HomeScreen = (props) => {
-  return (
-  <Container>
-    <StatusBar hidden={true}/>
-    <Text>Welcome home</Text>
-    <Button
-      title="What time is it?"
-      onPress={() => props.navigation.navigate("Time")}
-    />
-    <Button
-      title="Press butan"
-      onPress={() => props.navigation.navigate("Count")}
-    />
-  </Container>
-  )
-}
-
-const TimeScreen = (props) => {
-  return(
-  <Container>
-    <StatusBar hidden={true}/>
-    <Text>The time is {Date.now()}</Text>
-    <Button
-      title="Ok great thanks"
-      onPress={() => props.navigation.navigate("Home")}
-    />
-  </Container>
-  )
 }
 
 const RootStack = createStackNavigator(
