+++ screens/Container.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+import { View } from 'react-native'
+
+const Container = styled(View)`
+  align-items: center;
+  background-color: pink;
+  display: flex;
+  justify-content: center;
+  height: 100%;
+  width: 100%;
+`
+
+export default Container
